feat: add calldata function
