+++ seth/src/lib.rs
@@ -3,6 +3,10 @@
 //! TODO
 use chrono::NaiveDateTime;
 use ethers_core::{
+    abi::{
+        token::{LenientTokenizer, StrictTokenizer, Tokenizer},
+        AbiParser, ParamType, Token,
+    },
     types::*,
     utils::{self, keccak256},
 };
@@ -12,6 +16,7 @@ use rustc_hex::{FromHexIter, ToHex};
 use std::str::FromStr;
 
 use dapp_utils::{encode_args, get_func, to_table};
+use eyre::WrapErr;
 
 // TODO: SethContract with common contract initializers? Same for SethProviders?
 
@@ -523,8 +528,70 @@ impl SimpleSeth {
         let namehash: String = node.to_hex();
         Ok(format!("0x{}", namehash))
     }
+
+    /// Parses string input as Token against the expected ParamType
+    pub fn parse_tokens(params: &[(ParamType, &str)], lenient: bool) -> eyre::Result<Vec<Token>> {
+        params
+            .iter()
+            .map(|&(ref param, value)| {
+                if lenient {
+                    LenientTokenizer::tokenize(param, value)
+                } else {
+                    StrictTokenizer::tokenize(param, value)
+                }
+            })
+            .collect::<Result<_, _>>()
+            .wrap_err("Failed to parse tokens")
+    }
+
+    /// Performs ABI encoding to produce the hexadecimal calldata with the given arguments.
+    ///
+    /// ```
+    /// # use seth::SimpleSeth as Seth;
+    ///
+    /// # fn main() -> eyre::Result<()> {
+    ///     assert_eq!(
+    ///         "0xb3de648b0000000000000000000000000000000000000000000000000000000000000001",
+    ///         Seth::calldata("f(uint a)", &["1"]).unwrap().as_str()
+    ///     );
+    /// #    Ok(())
+    /// # }
+    /// ```
+    pub fn calldata(sig: impl AsRef<str>, args: &[impl AsRef<str>]) -> Result<String> {
+        let fun = AbiParser::default().parse_function(sig.as_ref())?;
+        let params: Vec<_> = fun
+            .inputs
+            .iter()
+            .map(|param| param.kind.clone())
+            .zip(args.iter().map(AsRef::as_ref))
+            .collect();
+        let tokens = SimpleSeth::parse_tokens(&params, true)?;
+        let calldata = fun.encode_input(&tokens)?;
+        Ok(format!("0x{}", calldata.to_hex::<String>()))
+    }
 }
 
 fn strip_0x(s: &str) -> &str {
     s.strip_prefix("0x").unwrap_or(s)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::SimpleSeth as Seth;
+
+    #[test]
+    fn calldata_uint() {
+        assert_eq!(
+            "0xb3de648b0000000000000000000000000000000000000000000000000000000000000001",
+            Seth::calldata("f(uint a)", &["1"]).unwrap().as_str()
+        );
+    }
+
+    #[test]
+    fn calldata_bool() {
+        assert_eq!(
+            "0x6fae94120000000000000000000000000000000000000000000000000000000000000000",
+            Seth::calldata("bar(bool)", &["false"]).unwrap().as_str()
+        );
+    }
 }
